requirements.txt
datetime
apache-airflow
pandas
requests
bs4 
re